+++ lib/src/painter/clock_painter.dart
@@ -14,15 +14,38 @@ class ClockPainter extends CustomPainter {
   double minAngle = 0;
   double maxAngle = 360;
 
-  final int maxCount;
-  final int currentCount;
+  final List<String> hourLabelList = [
+    "12",
+    "1",
+    "2",
+    "3",
+    "4",
+    "5",
+    "6",
+    "7",
+    "8",
+    "9",
+    "10",
+    "11",
+  ];
+
+  final int hourMaxCount;
+  final int hourCurrentCount;
+  final int minuteMaxCount;
+  final int minuteCurrentCount;
+  final int secondMaxCount;
+  final int secondCurrentCount;
 
   ClockPainter({
     required this.interval,
     required this.lineCount,
     required this.color,
-    required this.maxCount,
-    required this.currentCount,
+    required this.secondMaxCount,
+    required this.secondCurrentCount,
+    required this.hourMaxCount,
+    required this.hourCurrentCount,
+    required this.minuteMaxCount,
+    required this.minuteCurrentCount,
   });
 
   @override
@@ -38,7 +61,8 @@ class ClockPainter extends CustomPainter {
       ..strokeWidth = halfPaintWidth;
 
     final Rect canvasBorderRect = Rect.fromLTRB(0, 0, size.width, size.height);
-    final double diameter = math.min(canvasBorderRect.width, canvasBorderRect.height);
+    final double diameter =
+        math.min(canvasBorderRect.width, canvasBorderRect.height);
     final double radius = diameter / 2;
     final Offset centerOffset = Offset(radius, radius);
 
@@ -66,8 +90,26 @@ class ClockPainter extends CustomPainter {
       centerOffset: centerOffset,
       radius: radius,
       canvasBorderRect: canvasBorderRect,
-      maxCount: maxCount,
-      currentCount: currentCount,
+      maxCount: hourMaxCount,
+      currentCount: hourCurrentCount,
+    );
+
+    _drawMinutePointer(
+      canvas: canvas,
+      centerOffset: centerOffset,
+      radius: radius,
+      canvasBorderRect: canvasBorderRect,
+      maxCount: minuteMaxCount,
+      currentCount: minuteCurrentCount,
+    );
+
+    _drawSecondPointer(
+      canvas: canvas,
+      centerOffset: centerOffset,
+      radius: radius,
+      canvasBorderRect: canvasBorderRect,
+      maxCount: secondMaxCount,
+      currentCount: secondCurrentCount,
     );
 
     _drawText(
@@ -179,8 +221,8 @@ class ClockPainter extends CustomPainter {
     required int maxCount,
     required int currentCount,
   }) {
-    final double hourPointerLength = radius / 2;
-    final double hourPointerStrokeWidth = hourPointerLength / 10;
+    final double hourPointerLength = radius / 3;
+    final double hourPointerStrokeWidth = radius / (3 * 10);
     Paint paint = Paint();
     paint
       ..color = Colors.green
@@ -191,6 +233,59 @@ class ClockPainter extends CustomPainter {
 
     canvas.save();
     canvas.translate(radius, radius);
+    canvas.rotate(_toAngleValue(-90));
+    canvas.rotate(_toAngleValue(sweep));
+    canvas.drawLine(Offset.zero, Offset(0 + hourPointerLength, 0), paint);
+    canvas.restore();
+  }
+
+  _drawMinutePointer({
+    required Canvas canvas,
+    required Offset centerOffset,
+    required double radius,
+    required Rect canvasBorderRect,
+    required int maxCount,
+    required int currentCount,
+  }) {
+    final double hourPointerLength = radius / 2;
+    final double hourPointerStrokeWidth = radius / (2 * 10);
+    Paint paint = Paint();
+    paint
+      ..color = Colors.yellow
+      ..strokeCap = StrokeCap.round
+      ..strokeWidth = hourPointerStrokeWidth;
+
+    double sweep = 360 / maxCount * currentCount;
+
+    canvas.save();
+    canvas.translate(radius, radius);
+    canvas.rotate(_toAngleValue(-90));
+    canvas.rotate(_toAngleValue(sweep));
+    canvas.drawLine(Offset.zero, Offset(0 + hourPointerLength, 0), paint);
+    canvas.restore();
+  }
+
+  _drawSecondPointer({
+    required Canvas canvas,
+    required Offset centerOffset,
+    required double radius,
+    required Rect canvasBorderRect,
+    required int maxCount,
+    required int currentCount,
+  }) {
+    final double hourPointerLength = radius / 3 * 2;
+    final double hourPointerStrokeWidth = radius / (1 * hourPointerLength);
+    Paint paint = Paint();
+    paint
+      ..color = Colors.red
+      ..strokeCap = StrokeCap.round
+      ..strokeWidth = hourPointerStrokeWidth;
+
+    double sweep = 360 / maxCount * currentCount;
+
+    canvas.save();
+    canvas.translate(radius, radius);
+    canvas.rotate(_toAngleValue(-90));
     canvas.rotate(_toAngleValue(sweep));
     canvas.drawLine(Offset.zero, Offset(0 + hourPointerLength, 0), paint);
     canvas.restore();
@@ -202,26 +297,40 @@ class ClockPainter extends CustomPainter {
     required double radius,
     required Rect canvasBorderRect,
   }) {
-    var textPainter = TextPainter(
-      text: const TextSpan(
-          text: "12",
-          style: TextStyle(
-            color: Colors.red,
-          )),
-      textDirection: TextDirection.ltr,
-    );
+    for (int index = 0, size = hourLabelList.length; index < size; index++) {
+      final double angle = 30.0 * index;
+      final String label = hourLabelList.elementAt(index);
 
-    textPainter.layout(
-      minWidth: 0,
-      maxWidth: canvasBorderRect.width,
-    );
+      var textPainter = TextPainter(
+        text: TextSpan(
+            text: label,
+            style: TextStyle(
+              color: Colors.grey,
+            )),
+        textDirection: TextDirection.ltr,
+      );
 
-    canvas.save();
-    canvas.translate(0, 0);
-    canvas.rotate(_toAngleValue(0));
-    canvas.translate(0, -(radius / 2));
-    textPainter.paint(canvas, centerOffset);
-    canvas.restore();
+      textPainter.layout(
+        minWidth: 0,
+        maxWidth: canvasBorderRect.width,
+      );
+
+      final double translateDistance = radius - fullPaintWidth;
+      final double textHalfWidth = textPainter.width / 2;
+      final double textHalfHeight = textPainter.height / 2;
+
+      canvas.save();
+      canvas.translate(radius, radius);
+      canvas.rotate(_toAngleValue(angle));
+      Offset offset =
+          Offset(0 - textHalfWidth, 0 - textHalfHeight - translateDistance);
+      canvas.translate(offset.dx, offset.dy);
+      canvas.translate(textHalfWidth, textHalfHeight);
+      canvas.rotate(_toAngleValue(-angle));
+      canvas.translate(-textHalfWidth, -textHalfHeight);
+      textPainter.paint(canvas, const Offset(0, 0));
+      canvas.restore();
+    }
   }
 
   @override