+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+## [1.0.0]
 
-* TODO: Describe initial release.
+* add: hour, minute, second pointer.
+
+![clock gif](document/gif/clock_gif.gif)
+
+## [0.0.1]
+
+* add: Describe initial release.